+++ app/config/config.testnet.js
@@ -1,6 +1,7 @@
 export default {
   testnet: true,
   derivationPath: "m/44'/1'/",
+  passwordMinLength: 3,
   adsRpcHost: 'https://rpc.e11.click/',
   operatorUrl: 'https://operator.e11.click/',
   vaultStorageKey: 'testnet-vault',