+++ app/utils/storage.js
@@ -1,8 +1,28 @@
 import config from '../config/config';
 
+function prepareForms(dirty) {
+  const forms = {};
+  Object.entries(dirty).forEach(([fi, form]) => {
+    const inputs = {};
+    Object.entries(form.inputs).forEach(([ii, input]) => {
+      inputs[ii] = {
+        ...input,
+        value: input.private ? '' : input.value,
+      };
+    });
+    forms[fi] = {
+      ...form,
+      inputs,
+    };
+  });
+  return forms;
+}
+
 function saveState(state) {
   chrome.storage.local.set({
     [config.routerStorageKey]: JSON.stringify(state.router),
+    [config.formsStorageKey]: JSON.stringify(prepareForms(state.pages)),
+    [config.transactionsStorageKey]: JSON.stringify(prepareForms(state.transactions)),
   });
   if (state.vault.selectedAccount) {
     chrome.storage.local.set({