+++ chrome/extension/adswallet.js
@@ -5,16 +5,22 @@ import BgClient from '../../app/utils/background';
 import { reload as reloadQueue } from '../../app/actions/queueActions';
 import './adswallet.css';
 
-
-function renderDOM(Root, initialState, config) {
+function createHistory(initialState) {
   const history = createHashHistory();
-  if (initialState.router && initialState.router.location) {
-    if (history.location.path !== initialState.router.location.path ||
-      history.location.search !== initialState.router.location.search ||
-      history.location.hash !== initialState.router.location.hash) {
+  if (history.location.pathname === '/') {
+    if (history.location.pathname !== initialState.router.location.pathname ||
+        history.location.search !== initialState.router.location.search ||
+        history.location.hash !== initialState.router.location.hash) {
       history.push(initialState.router.location);
     }
   }
+
+  return history;
+}
+
+
+function renderDOM(Root, initialState, config) {
+  const history = createHistory(initialState);
   const createStore = require('../../app/store/configureStore');
 
   const store = createStore(initialState, history);
@@ -45,7 +51,9 @@ BgClient.getSession((session) => {
   chrome.storage.local.get([
     config.routerStorageKey,
     config.queueStorageKey,
-    config.accountStorageKey
+    config.accountStorageKey,
+    config.formsStorageKey,
+    config.transactionsStorageKey,
   ], (obj) => {
     const initialState = {};
     if (obj[config.routerStorageKey]) {
@@ -53,6 +61,12 @@ BgClient.getSession((session) => {
     }
     if (obj[config.queueStorageKey]) {
       initialState.queue = JSON.parse(obj[config.queueStorageKey]);
+    }
+    if (obj[config.formsStorageKey]) {
+      initialState.pages = JSON.parse(obj[config.formsStorageKey]);
+    }
+    if (obj[config.transactionsStorageKey]) {
+      initialState.transactions = JSON.parse(obj[config.transactionsStorageKey]);
     }
     VaultCrypt.load((vault) => {
       initialState.vault = vault;