+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2019-01-11
+### Added
+- Store form temporary state
+### Changed
+- Password length set to 8 chars
+### Fixed
+- Location preservation
+
 ## [0.1.0] - 2019-01-10
 ### Added
 - Creating transactions
@@ -18,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Storage encryption
 - Docs
 
-[Unreleased]: https://github.com/adshares/ads-browser-wallet/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/adshares/ads-browser-wallet/compare/v0.1.1...HEAD
 
+[0.1.1]: https://github.com/adshares/ads-browser-wallet/releases/tag/v0.1.1
 [0.1.0]: https://github.com/adshares/ads-browser-wallet/releases/tag/v0.1.0